fix: перенесена часть скрипта с catalog.js -> pers
