+++ personalaccount.html
@@ -34,6 +34,6 @@
         </div>
       </dialog>
     </div>
-    <script src="./src/personalaccount.js"></script>
+    <script type="module" src="./src/personalaccount.js"></script>
   </body>
 </html>
